I changed the password_reset to email_action_handler, the email_action_handler handles all email actions like password reset and email verification
